+++ src/main/java/dev/exonbyte/student_organizer/controllers/StudentController.java
@@ -1,57 +1,39 @@
 package dev.exonbyte.student_organizer.controllers;
 
-import dev.exonbyte.student_organizer.DTO.StudentDTO;
 import dev.exonbyte.student_organizer.DTO.StudentResponse;
 import dev.exonbyte.student_organizer.services.StudentServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
-@RestController
-@RequestMapping("/api/")
+import java.util.Optional;
+
+
+@Controller
 public class StudentController {
 
-    private final StudentServiceImp studentService;
-
     @Autowired
-    public StudentController(StudentServiceImp studentService) {
-        this.studentService = studentService;
+    private StudentServiceImp studentService;
+
+    @RequestMapping(value = "/students", method = RequestMethod.GET)
+    public String showAllStudents(Model model,
+                                  @RequestParam(value = "pageNumber", defaultValue = "0", required = false) Optional<Integer> page,
+                                  @RequestParam(value = "pageSize", defaultValue = "5", required = false) Optional<Integer> size
+    ) {
+        int pageNumber = page.orElse(0);
+        int pageSize = size.orElse(10);
+
+        StudentResponse students = studentService.getAllStudents(pageNumber, pageSize);
+
+        model.addAttribute("students", students);
+        model.addAttribute("pageSize", students.getPageSize());
+        model.addAttribute("pageNumber", students.getPageNumber());
+
+        return "students";
     }
 
-    @GetMapping("/students")
-    public ResponseEntity<StudentResponse> getAllStudents(
-            @RequestParam(value = "pageNumber", defaultValue = "0", required = false) int pageNumber,
-            @RequestParam(value = "pageSize", defaultValue = "10", required = false) int pageSize
-    )
-    {
-        return new ResponseEntity<>(studentService.getAllStudents(pageNumber, pageSize), HttpStatus.OK);
-    }
 
-    @GetMapping("students/{id}")
-    public ResponseEntity<StudentDTO> studentDetail(@PathVariable Integer id) {
-        return new ResponseEntity<>(studentService.getStudentById(id), HttpStatus.OK);
-    }
-
-    @PostMapping("students/create")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<StudentDTO> createStudent(@RequestBody StudentDTO studentDTO) {
-        return new ResponseEntity<>(studentService.createStudent(studentDTO), HttpStatus.CREATED);
-    }
-
-    @PutMapping("students/{id}/update")
-    public ResponseEntity<StudentDTO> updateStudent(@PathVariable Integer id, @RequestBody StudentDTO student) {
-        studentService.updateStudent(id,student);
-        return new ResponseEntity<>(student, HttpStatus.OK);
-    }
-
-    @DeleteMapping("students/{id}/delete")
-    public void deleteStudent(@PathVariable Integer id) {
-        studentService.deleteStudentById(id);
-    }
-
-    @DeleteMapping("students/purge")
-    public void deleteAllStudents(){
-        studentService.deleteAllStudents();
-    }
 }
